[JUV-107] 프로젝트 폴더 구조 세미 개선하기 (#21) / (#22)
