--- Server/app/business/service/user/UserAbstract.py
@@ -1,3 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-
