+++ Server/app/core/compact/CompactConfig.py
@@ -1,4 +1,6 @@
 from dataclasses import dataclass
+from typing import Union
+from Server.app.core.ConfigAbstract import ConfigClass, CompactConfigClass, TemplateClass
 from Server.app.core.configs.BaseConfig import BaseConfig
 from Server.app.core.configs.ClovaConfig import ClovaConfig
 from Server.app.core.configs.DBConfig import DBConfig
@@ -10,13 +12,13 @@ from Server.app.core.configs.SlackTemplates import SlackTemplates
 
 @dataclass(frozen=False)
 class CompactConfig:
-    BASE: BaseConfig = BaseConfig()
-    CLOVA: ClovaConfig = ClovaConfig()
-    DB: DBConfig = DBConfig()
-    USER: UserConfig = UserConfig()
-    SLACK: SlackConfig = SlackConfig()
-    TEAMS: TeamsConfig = TeamsConfig()
-    SLACK_TEMPLATES: SlackTemplates = SlackTemplates()
+    BASE: Union[ConfigClass, BaseConfig] = BaseConfig()
+    CLOVA: Union[ConfigClass, ClovaConfig] = ClovaConfig()
+    DB: Union[ConfigClass, DBConfig] = DBConfig()
+    USER: Union[ConfigClass, UserConfig] = UserConfig()
+    SLACK: Union[ConfigClass, SlackConfig] = SlackConfig()
+    TEAMS: Union[ConfigClass, TeamsConfig] = TeamsConfig()
+    SLACK_TEMPLATES: Union[TemplateClass, SlackTemplates] = SlackTemplates()
 
 
-CONFIG: CompactConfig = CompactConfig()
+CONFIG: Union[CompactConfigClass, CompactConfig] = CompactConfig()
