+++ Server/app/core/configs/BaseConfig.py
@@ -1,10 +1,11 @@
 import os
 from dotenv import load_dotenv
 from dataclasses import dataclass
+from Server.app.core.ConfigAbstract import ConfigClass
 
 
 @dataclass(frozen=True)
-class BaseConfig(object):
+class BaseConfig(ConfigClass):
     load_dotenv(
         dotenv_path="./.env",
         verbose=True
