+++ Server/app/core/configs/SlackConfig.py
@@ -1,8 +1,9 @@
 import os
 from dataclasses import dataclass
+from Server.app.core.ConfigAbstract import ConfigClass
 
 
 @dataclass(frozen=True)
-class SlackConfig:
+class SlackConfig(ConfigClass):
     TOKEN: str = os.getenv("SLACK_TOKEN")
