+++ Server/app/core/configs/SlackTemplates.py
@@ -1,8 +1,9 @@
 from dataclasses import dataclass, field
+from Server.app.core.ConfigAbstract import TemplateClass
 
 
 @dataclass(frozen=False)
-class SlackTemplates:
+class SlackTemplates(TemplateClass):
     TEST_TEMPLATES: dict = field(default_factory=lambda: dict(
         test_template=[
             {
