+++ Server/app/core/configs/TeamsConfig.py
@@ -1,7 +1,8 @@
 import os
 from dataclasses import dataclass
+from Server.app.core.ConfigAbstract import ConfigClass
 
 
 @dataclass(frozen=True)
-class TeamsConfig:
+class TeamsConfig(ConfigClass):
     pass
