+++ Server/app/core/configs/UserConfig.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
+from Server.app.core.ConfigAbstract import ConfigClass
 
 
 @dataclass(frozen=True)
-class UserConfig:
+class UserConfig(ConfigClass):
     pass
