--- Server/app/outer/connection/ai/AiAbstract.py
@@ -1,1 +1,0 @@
-from abc import ABCMeta, abstractmethod
