--- Server/app/outer/connection/clova/ClovaAbstract.py
@@ -1,15 +1,0 @@
-from abc import ABCMeta, abstractmethod
-from Server.app.outer.connection.clova.ClovaTextDTO import ClovaTextDTO
-
-
-class ClovaTextOcrModel(metaclass=ABCMeta):
-    @abstractmethod
-    def request(self) -> ClovaTextDTO:
-        pass
-
-
-class ClovaRequestModel(metaclass=ABCMeta):
-    @abstractmethod
-    def request_to_clova(self) -> None:
-        pass
-
