--- Server/app/outer/connection/clova/ClovaTextDTO.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass(frozen=False)
-class ClovaTextDTO:
-    text: str
-    requestStatusCode: int
