--- Server/app/outer/db/ai/AiAbstract.py
@@ -1,1 +1,0 @@
-from abc import ABCMeta, abstractmethod
