--- Server/app/outer/db/base/BaseDBAbstract.py
@@ -1,62 +1,0 @@
-from abc import ABCMeta, abstractmethod
-from typing import List
-
-
-class BaseDBSelectModel(metaclass=ABCMeta):
-    @abstractmethod
-    def connect(self) -> None:
-        pass
-
-    @abstractmethod
-    def single_select(self, query: str) -> str:
-        pass
-
-    @abstractmethod
-    def multi_select(self, query: str) -> List[str]:
-        pass
-
-
-class BaseDBInsertModel(metaclass=ABCMeta):
-    @abstractmethod
-    def connect(self) -> None:
-        pass
-
-    @abstractmethod
-    def single_insert(self) -> bool:
-        pass
-
-    @abstractmethod
-    def multi_insert(self) -> bool:
-        pass
-
-
-class BaseDBUpdateModel(metaclass=ABCMeta):
-    @abstractmethod
-    def connect(self) -> None:
-        pass
-
-    @abstractmethod
-    def single_update(self) -> bool:
-        pass
-
-    @abstractmethod
-    def multi_update(self) -> bool:
-        pass
-
-
-class BaseDBDeleteModel(metaclass=ABCMeta):
-    @abstractmethod
-    def connect(self) -> None:
-        pass
-
-    @abstractmethod
-    def single_delete(self) -> bool:
-        pass
-
-    @abstractmethod
-    def multi_delete(self) -> bool:
-        pass
-
-    @abstractmethod
-    def truncate(self) -> bool:
-        pass
