--- Server/app/outer/db/clova/ClovaAbstract.py
@@ -1,1 +1,0 @@
-from abc import ABCMeta, abstractmethod
